Renamed: component PickerPageSize.vue
