+++ apps/frontend/src/components/page-products/PickerPageSize.vue
